Decided to use a TDD approach for implementing the AnnuityRepaymentScheduleCalculator and completed some WebMvc testing for the controller methods

+++ src/main/java/com/aledsdavies/ldms/models/RepaymentPeriod.java
@@ -12,8 +12,8 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class RepaymentPeriod {
     private int period;
-    private BigDecimal payment;
-    private BigDecimal principle;
-    private BigDecimal interest;
-    private BigDecimal balance;
+    private double payment;
+    private double principle;
+    private double interest;
+    private double balance;
 }
